new wasm with csp primitives to generate and verify proofs
