--- cmd/ballotproof-wasm/nodejstest/Makefile
@@ -1,9 +1,0 @@
-.PHONY: all build start
-
-all: build
-
-build:
-	@echo "Building WASM binary for Node.js..."
-	cd .. && GOGC=off GOOS=js GOARCH=wasm go build -ldflags="-s -w" -o nodejstest/ballotproof.wasm .
-	@echo "Starting Node.js application..."
-	node index.js
